+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # Notas de versão
 - Atualizado CACERT
+- Adicionado modo Multithreading